Split lammps.py into core.py and pylammps.py
